wrote activation_diff

+++ src/kernels-glsl/activation.glsl
@@ -15,24 +15,24 @@ layout(binding = 1, std430) buffer c_buf { dtype c[]; };
 
 layout(push_constant, std430) uniform activation
 {
-	uint64_t size;
+	uint size;
 #if USE_BDA
 	__global dtype *a;
 #endif
-	uint64_t a_offset;
+	uint a_offset;
 #if USE_BDA
 	__global dtype *c,
 #endif
-	uint64_t c_offset;
+	uint c_offset;
 };
 
 void main()
 {
-    uint64_t pos = get_global_id(0);
+    uint pos = get_global_id(0);
     if(pos >= size)
         return;
-	uint64_t a_pos = pos + a_offset;
-	uint64_t c_pos = pos + c_offset;
+	uint a_pos = pos + a_offset;
+	uint c_pos = pos + c_offset;
 	dtype a_val = a[a_pos];
     c[c_pos] = ACTIVATION_F(a_val);
 }
